requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-telegram-bot==20.7
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.12.0
pydantic==2.5.3
pydantic-settings==2.1.0
click==8.1.7
httpx==0.26.0
python-dateutil==2.8.2
pgvector==0.2.4
schedule==1.2.0
pytest==7.4.4
pytest-asyncio==0.23.3

